[iris] Merge CoreWeave network setup into one idempotent install_cw_network.py
Fold install_federation_ingress.py into install_traefik_proxy.py and rename the
result install_cw_network.py: one operator script per CoreWeave cluster that
installs everything the controller can't do itself — Traefik, cert-manager, the
HTTP-01 ClusterIssuers, and the IP-locked federation ingress — in a single
idempotent pass.

Two scripts meant an operator could run the stack bootstrap but forget the
federation route (or vice versa); they also sat at the same layer (operator
kubeconfig, CRD-level installs the controller ServiceAccount can't do), so the
split bought nothing. `install` now takes both --acme-email and --allow-source
and applies the whole stack; --skip-traefik/--skip-cert-manager/--skip-issuers
reconcile just the federation route on a cluster that already has the stack.
`uninstall` tears it all down (federation route included) and verifies.

The federation route keeps its behavior: reuse the controller's iris-cw-* host,
reuse the existing /proxy cert where one exists (else issue a fresh one), and
warn loudly if the target controller is still permissive.
