+++ lib/iris/scripts/install_cw_network.py
@@ -1,0 +1,962 @@
+#!/usr/bin/env python3
+# Copyright The Marin Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Set up everything a CoreWeave (CKS) controller needs on the network that the controller can't install itself.
+
+CKS ships no ingress controller and no TLS issuer, and the controller's own
+ServiceAccount can't install CRDs, so an operator runs this once per cluster from
+their kubeconfig. It installs, idempotently, the whole networking stack in one
+pass so there is no "ran one script but not the other" gap:
+
+  * **Traefik** — the ingress controller; CKS gives its LoadBalancer a wildcard
+    ``*.coreweave.app`` DNS name.
+  * **cert-manager** — issues the TLS certs Traefik terminates.
+  * **HTTP-01 Let's Encrypt ClusterIssuers** — CoreWeave's bundled issuers use
+    DNS-01 against ``acme.coreweave.com`` and only cover ``*.coreweave.app``; a
+    custom host (``iris-cw-<cluster>.oa.dev`` CNAME'd to the coreweave.app FQDN)
+    needs HTTP-01, which validates through Traefik for any host that resolves to
+    the LoadBalancer.
+  * **The federation ingress** — an IP-locked route publishing the controller's
+    RPC surface to the GCP ``marin`` controller (see below).
+
+The controller creates its own ``/proxy`` Ingress at start-up
+(``K8sControllerProvider.start_controller``); this script installs the cluster-wide
+prerequisites that Ingress depends on, and adds the federation route beside it.
+
+Federation ingress — how the GCP ``marin`` controller reaches this one:
+
+    marin federates whole jobs to the CoreWeave controllers by dialing their RPC
+    surface directly (the pull model — CoreWeave must be reachable *inbound*).
+    CoreWeave has no user surface of its own: end users reach Iris only through
+    ``iris.oa.dev`` (IAP), and marin federates outward, so the only external caller
+    of a CoreWeave controller is the marin controller. That makes the auth surface
+    two factors, both required and neither alone sufficient:
+
+      1. **IP allowlist** — a Traefik ``ipAllowList`` Middleware admits only the
+         marin controller's egress IP. The *network* factor.
+      2. **The controller's own auth** — the method-scoped ``aud="federation"``
+         verifier gates the handoff RPCs, and the general auth chain gates the
+         rest. The *identity* factor. The ingress does NOT do method policy.
+
+    Because the controller is the identity gate, it MUST be enforcing. A permissive
+    (null-auth) controller behind only an IP lock hands anonymous admin over its
+    entire control plane to anything arriving from the allowlisted IP — so this
+    script warns loudly when its target is still permissive. ``cw-us-east-02a`` is
+    enforcing (``trusted_cidrs``); ``cw-rno2a`` gains ``trusted_cidrs`` too (see
+    ``docs/coreweave.md``).
+
+    The route reuses the controller's ``iris-cw-*.oa.dev`` host rather than a
+    dedicated name. Where the cluster already publishes ``/proxy``
+    (``cw-us-east-02a``), that world-open route keeps its own more-specific Traefik
+    router (Traefik prefers the longer path match), so ``/proxy`` stays open while
+    the federation route allowlists the rest of the surface and reuses the existing
+    TLS cert. Where there is no ``/proxy`` host (``cw-rno2a``), the federation route
+    publishes the whole controller host and cert-manager issues a fresh cert.
+
+Target a cluster with ``--cluster <name>`` (resolved the same way ``iris --cluster``
+does): the host, ingress class, namespace, controller Service, and kubeconfig all
+come from that config's ``controller.coreweave`` block, so they match what the
+controller itself uses.
+
+Without ``--apply`` every subcommand only prints its plan. ``uninstall`` tears the
+whole stack back down — the federation route, the helm releases, their namespaces,
+and the cluster-scoped CRDs/webhooks/RBAC/IngressClass a namespace delete would
+orphan — then verifies nothing remains.
+
+Usage:
+    # Dry-run (default): prints the helm commands, manifests, and pre-flight findings.
+    uv run lib/iris/scripts/install_cw_network.py --cluster cw-rno2a \\
+        install --acme-email you@oa.dev --allow-source 203.0.113.7
+    # Apply the whole stack, then CNAME the host to the Traefik LoadBalancer FQDN:
+    uv run lib/iris/scripts/install_cw_network.py --cluster cw-rno2a \\
+        install --acme-email you@oa.dev --allow-source 203.0.113.7 --apply
+    # Re-run to converge just the federation route (skip the already-installed stack):
+    uv run lib/iris/scripts/install_cw_network.py --cluster cw-rno2a install \\
+        --allow-source 203.0.113.7 --skip-traefik --skip-cert-manager --skip-issuers \\
+        --cluster-issuer letsencrypt-http01-prod --apply
+    # Tear the whole stack down and verify nothing remains:
+    uv run lib/iris/scripts/install_cw_network.py --cluster cw-rno2a uninstall --apply
+"""
+
+import ipaddress
+import subprocess
+import time
+from typing import NamedTuple
+
+import click
+import yaml
+from iris.cli.connect import IRIS_CLUSTER_CONFIG_DIRS
+from iris.cluster.config import load_config
+from rigging.config_discovery import resolve_cluster_config
+
+DEFAULT_INGRESS_CLASS = "traefik"
+
+CW_REPO_NAME = "coreweave"
+CW_REPO_URL = "https://charts.core-services.ingress.coreweave.com"
+TRAEFIK_CHART = f"{CW_REPO_NAME}/traefik"
+CERT_MANAGER_CHART = f"{CW_REPO_NAME}/cert-manager"
+
+LE_ACME = {
+    "prod": "https://acme-v02.api.letsencrypt.org/directory",
+    "staging": "https://acme-staging-v02.api.letsencrypt.org/directory",
+}
+# ClusterIssuer names this script creates; use as controller.coreweave.cluster_issuer.
+ISSUER_NAMES = {"prod": "letsencrypt-http01-prod", "staging": "letsencrypt-http01-staging"}
+
+CLUSTERISSUER_CRD = "clusterissuers.cert-manager.io"
+_CRD_WAIT_SECONDS = 120.0
+
+# API groups whose CRDs the charts register; a namespace delete orphans these.
+CRD_GROUPS = ("cert-manager.io", "traefik.io")
+# Cluster-scoped kinds the charts create that `helm uninstall` normally removes
+# but a namespace-only teardown leaves behind.
+SWEEP_KINDS = (
+    "validatingwebhookconfiguration",
+    "mutatingwebhookconfiguration",
+    "clusterrole",
+    "clusterrolebinding",
+)
+INSTANCE_LABEL = "app.kubernetes.io/instance"
+
+# CoreWeave's External Hostname Controller allocates the LB's *.coreweave.app
+# FQDN asynchronously and reports it in the Service's ExternalRecords condition.
+_FQDN_JSONPATH = '{.status.conditions[?(@.type=="ExternalRecords")].message}'
+_FQDN_WAIT_SECONDS = 90.0
+_COREWEAVE_APP = ".coreweave.app"
+
+# Federation route: an IP-locked Ingress + ipAllowList Middleware on the controller host.
+_INGRESS_NAME = "iris-federation"
+_MIDDLEWARE_NAME = "iris-federation-ipallowlist"
+_MIDDLEWARE_CRD = "middlewares.traefik.io"
+# TLS secret for the federation route when the cluster has no /proxy cert to reuse
+# (e.g. cw-rno2a, which publishes no /proxy host today).
+DEFAULT_TLS_SECRET = "iris-controller-fed-tls"
+# Staging first to avoid Let's Encrypt rate limits while DNS/allowlist are shaken
+# out; flip to the prod issuer once the staging cert validates.
+DEFAULT_CLUSTER_ISSUER = ISSUER_NAMES["staging"]
+# _auth_mode's value for a permissive controller (no login arm, no trusted_cidrs).
+_NULL_AUTH = "null-auth"
+
+
+# --------------------------------------------------------------------------
+# Thin I/O helpers (subprocess via arg lists — never shell=True).
+# --------------------------------------------------------------------------
+def helm_flags(kubeconfig: str | None, context: str | None) -> list[str]:
+    flags: list[str] = []
+    if kubeconfig:
+        flags += ["--kubeconfig", kubeconfig]
+    if context:
+        flags += ["--kube-context", context]
+    return flags
+
+
+def kubectl_flags(kubeconfig: str | None, context: str | None) -> list[str]:
+    flags: list[str] = []
+    if kubeconfig:
+        flags += ["--kubeconfig", kubeconfig]
+    if context:
+        flags += ["--context", context]
+    return flags
+
+
+def run(cmd: list[str], **kwargs) -> subprocess.CompletedProcess:
+    """Run a command from an arg list, echoing it first."""
+    click.secho(f"$ {' '.join(cmd)}", fg="bright_black")
+    return subprocess.run(cmd, **kwargs)
+
+
+def kubectl_apply_docs(docs: list[dict], kflags: list[str], what: str) -> None:
+    """Apply a list of manifests via ``kubectl apply -f -`` (stdin)."""
+    manifests = yaml.safe_dump_all(docs, default_flow_style=False, sort_keys=False)
+    proc = run(["kubectl", *kflags, "apply", "-f", "-"], input=manifests, text=True)
+    if proc.returncode != 0:
+        raise click.ClickException(f"kubectl apply of {what} failed")
+
+
+def resource_present(kind: str, name: str, kflags: list[str]) -> bool:
+    """True if ``kind/name`` exists; queries quietly, emitting no output."""
+    result = subprocess.run(
+        ["kubectl", *kflags, "get", kind, name],
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.DEVNULL,
+    )
+    return result.returncode == 0
+
+
+def wait_for_crd(crd: str, kflags: list[str]) -> None:
+    """Block until ``crd`` is established (cert-manager registers its CRDs async)."""
+    deadline = time.monotonic() + _CRD_WAIT_SECONDS
+    while time.monotonic() < deadline:
+        if resource_present("crd", crd, kflags):
+            return
+        time.sleep(3.0)
+    raise click.ClickException(f"CRD {crd} not present after {_CRD_WAIT_SECONDS:.0f}s (is cert-manager installed?)")
+
+
+def read_traefik_fqdn(release: str, namespace: str, kflags: list[str]) -> str:
+    """Poll the Traefik LoadBalancer Service for its allocated ``*.coreweave.app`` FQDN.
+
+    Returns the bare FQDN, or "" if not allocated within the wait window (the
+    External Hostname Controller assigns it asynchronously). Polls quietly (no
+    per-attempt command echo) since it may take several tries.
+    """
+    click.secho(f"==> Reading Traefik LoadBalancer FQDN (svc/{release} -n {namespace}) …", fg="blue", bold=True)
+    deadline = time.monotonic() + _FQDN_WAIT_SECONDS
+    while time.monotonic() < deadline:
+        result = subprocess.run(
+            ["kubectl", *kflags, "get", "svc", release, "-n", namespace, "-o", f"jsonpath={_FQDN_JSONPATH}"],
+            capture_output=True,
+            text=True,
+        )
+        for token in (result.stdout or "").replace(",", " ").split():
+            candidate = token.strip(".;\"'")
+            if candidate.endswith(_COREWEAVE_APP):
+                return candidate
+        time.sleep(3.0)
+    return ""
+
+
+# --------------------------------------------------------------------------
+# Settings, resolved once at the command group
+# --------------------------------------------------------------------------
+class NetworkSettings(NamedTuple):
+    """Target cluster, controller coordinates, and chart placement for the CW network stack."""
+
+    cluster: str
+    namespace: str  # controller namespace; the federation Ingress + Middleware live here
+    service_name: str  # controller Service the federation route backends onto
+    port: int
+    ingress_class: str
+    proxy_host: str  # the cluster's existing /proxy host, if any (host + cert to reuse)
+    proxy_tls_secret: str  # the cluster's existing /proxy TLS secret, if any
+    auth_mode: str  # "null-auth" (permissive) | "iap" | "cidr"
+    kubeconfig: str | None
+    context: str | None
+    traefik_namespace: str
+    traefik_release: str
+    cert_manager_namespace: str
+    cert_manager_release: str
+
+
+def _auth_mode(config) -> str:
+    """The controller's request-auth mode.
+
+    ``null-auth`` means the RPC surface is PERMISSIVE: no login-provider arm and
+    no ``trusted_cidrs``, so every request is admitted as the anonymous admin.
+    """
+    auth = config.auth
+    if auth is None or (auth.provider_kind() is None and not auth.trusted_cidrs):
+        return _NULL_AUTH
+    return auth.provider_kind() or "cidr"
+
+
+class _DerivedConfig(NamedTuple):
+    """Controller coordinates read from a named Iris cluster config."""
+
+    namespace: str
+    service_name: str
+    port: int
+    ingress_class: str
+    proxy_host: str
+    proxy_tls_secret: str
+    auth_mode: str
+    kubeconfig: str
+    context: str
+
+
+def _derive_from_cluster(name: str) -> _DerivedConfig:
+    """Read a named Iris cluster config and return its CoreWeave controller coordinates.
+
+    Resolves ``name`` the same way ``iris --cluster`` does, so the namespace,
+    controller Service, kubeconfig, and kube context match what the controller
+    itself uses.
+    """
+    try:
+        path = resolve_cluster_config(name, dirs=IRIS_CLUSTER_CONFIG_DIRS)
+    except FileNotFoundError as exc:
+        raise click.ClickException(f"Unknown cluster {name!r}; run `iris cluster list`.") from exc
+    config = load_config(str(path))
+    cw = config.controller.coreweave
+    if cw is None:
+        raise click.ClickException(f"Cluster {name!r} has no controller.coreweave block — not a CoreWeave cluster.")
+    platform = config.platform.coreweave
+    return _DerivedConfig(
+        namespace=(platform.namespace if platform else "") or "iris",
+        service_name=cw.service_name or "iris-controller-svc",
+        port=cw.port or 10000,
+        ingress_class=cw.ingress_class or DEFAULT_INGRESS_CLASS,
+        proxy_host=cw.public_proxy_host,
+        proxy_tls_secret=cw.tls_secret,
+        auth_mode=_auth_mode(config),
+        kubeconfig=(platform.kubeconfig_path if platform else "") or "",
+        context=(platform.kube_context if platform else "") or "",
+    )
+
+
+def _default_host(settings: NetworkSettings) -> str:
+    """Reuse the cluster's /proxy host, else ``iris-cw-<cluster>.oa.dev``."""
+    if settings.proxy_host:
+        return settings.proxy_host
+    short = settings.cluster[len("cw-") :] if settings.cluster.startswith("cw-") else settings.cluster
+    return f"iris-cw-{short}.oa.dev"
+
+
+def _normalize_source(value: str) -> str:
+    """Validate one allowlist entry and return it in CIDR form (bare IP -> /32,/128)."""
+    try:
+        if "/" in value:
+            ipaddress.ip_network(value, strict=False)
+            return value
+        ip = ipaddress.ip_address(value)
+    except ValueError as exc:
+        raise click.BadParameter(f"{value!r} is not a valid IP or CIDR: {exc}", param_hint="--allow-source") from exc
+    return f"{value}/{32 if ip.version == 4 else 128}"
+
+
+# --------------------------------------------------------------------------
+# Manifests
+# --------------------------------------------------------------------------
+def _http01_issuer(env: str, email: str, ingress_class: str) -> dict:
+    """A Let's Encrypt HTTP-01 ClusterIssuer validated through ``ingress_class``.
+
+    HTTP-01 (not CoreWeave's bundled DNS-01) so it can issue for a custom host,
+    which the coreweave.app DNS-01 webhook cannot. Requires the host to already
+    resolve to the Traefik LoadBalancer before issuance.
+    """
+    return {
+        "apiVersion": "cert-manager.io/v1",
+        "kind": "ClusterIssuer",
+        "metadata": {"name": ISSUER_NAMES[env]},
+        "spec": {
+            "acme": {
+                "server": LE_ACME[env],
+                "email": email,
+                "privateKeySecretRef": {"name": f"{ISSUER_NAMES[env]}-account-key"},
+                "solvers": [{"http01": {"ingress": {"ingressClassName": ingress_class}}}],
+            }
+        },
+    }
+
+
+def _build_ipallowlist_middleware(*, namespace: str, source_ranges: list[str], xff_depth: int) -> dict:
+    """A Traefik ``ipAllowList`` Middleware admitting only ``source_ranges``.
+
+    By default Traefik matches the client's direct transport peer (``RemoteAddr``).
+    If the CoreWeave LoadBalancer SNATs (so Traefik sees the LB, not the real
+    client), set ``xff_depth`` to the number of trusted proxy hops and Traefik
+    reads the client IP from ``X-Forwarded-For`` instead. Verify which applies by
+    testing a refused request from a non-allowlisted host.
+    """
+    ip_allow_list: dict = {"sourceRange": source_ranges}
+    if xff_depth > 0:
+        ip_allow_list["ipStrategy"] = {"depth": xff_depth}
+    return {
+        "apiVersion": "traefik.io/v1alpha1",
+        "kind": "Middleware",
+        "metadata": {"name": _MIDDLEWARE_NAME, "namespace": namespace},
+        "spec": {"ipAllowList": ip_allow_list},
+    }
+
+
+def _build_federation_ingress(
+    *,
+    namespace: str,
+    service_name: str,
+    port: int,
+    host: str,
+    ingress_class: str,
+    tls_secret: str,
+    cluster_issuer: str,
+) -> dict:
+    """Ingress routing the controller host to the RPC surface, IP-locked by the Middleware.
+
+    A single catch-all path attaches the ``ipAllowList`` Middleware via the
+    router-middlewares annotation (``<ns>-<name>@kubernetescrd``). Where the
+    cluster also publishes ``/proxy``, that route keeps its own more-specific
+    Traefik router (Traefik prefers the longer path match), so ``/proxy`` stays
+    world-open while everything else is allowlisted here.
+
+    ``cluster_issuer`` empty means the TLS cert is managed elsewhere (the ``/proxy``
+    Ingress) and this route only references ``tls_secret``; set it to have
+    cert-manager issue the cert here. cert-manager's HTTP-01 solver runs on its own
+    unrestricted Ingress, so the allowlist does not block ACME validation.
+    """
+    annotations = {
+        "traefik.ingress.kubernetes.io/router.middlewares": f"{namespace}-{_MIDDLEWARE_NAME}@kubernetescrd",
+    }
+    if cluster_issuer:
+        annotations["cert-manager.io/cluster-issuer"] = cluster_issuer
+    spec: dict = {
+        "ingressClassName": ingress_class,
+        "rules": [
+            {
+                "host": host,
+                "http": {
+                    "paths": [
+                        {
+                            "path": "/",
+                            "pathType": "Prefix",
+                            "backend": {"service": {"name": service_name, "port": {"number": port}}},
+                        }
+                    ]
+                },
+            }
+        ],
+    }
+    if tls_secret:
+        spec["tls"] = [{"hosts": [host], "secretName": tls_secret}]
+    return {
+        "apiVersion": "networking.k8s.io/v1",
+        "kind": "Ingress",
+        "metadata": {"name": _INGRESS_NAME, "namespace": namespace, "annotations": annotations},
+        "spec": spec,
+    }
+
+
+# --------------------------------------------------------------------------
+# Pre-flight
+# --------------------------------------------------------------------------
+def _warn_if_permissive(settings: NetworkSettings) -> None:
+    """Warn when the target controller admits unauthenticated callers.
+
+    With the federation route exposing the controller behind only an IP allowlist,
+    a permissive controller hands anonymous admin over its whole control plane to
+    anything from the allowlisted IP — so the controller must be enforcing before
+    that route is a real boundary.
+    """
+    if settings.auth_mode != _NULL_AUTH:
+        return
+    click.secho(
+        f"\nwarn: {settings.cluster} runs NULL-AUTH (permissive) — its RPC surface admits any caller\n"
+        "      as the anonymous admin. Behind only an IP allowlist that means the ENTIRE control\n"
+        "      plane is exposed to the allowlisted IP with no identity check. Make the controller\n"
+        "      enforcing first (add auth.trusted_cidrs, as cw-us-east-02a has) so an off-cluster\n"
+        "      request must present a valid bearer. See docs/coreweave.md.",
+        fg="yellow",
+        err=True,
+    )
+
+
+# --------------------------------------------------------------------------
+# Install
+# --------------------------------------------------------------------------
+def install(
+    settings: NetworkSettings,
+    *,
+    host: str,
+    source_ranges: list[str],
+    tls_secret: str,
+    cluster_issuer: str,
+    xff_depth: int,
+    acme_email: str | None,
+    traefik_version: str | None,
+    cert_manager_version: str | None,
+    skip_traefik: bool,
+    skip_cert_manager: bool,
+    skip_issuers: bool,
+    apply: bool,
+) -> None:
+    hflags = helm_flags(settings.kubeconfig, settings.context)
+    kflags = kubectl_flags(settings.kubeconfig, settings.context)
+
+    if not skip_issuers and not acme_email:
+        raise click.ClickException(
+            "--acme-email is required to create the HTTP-01 ClusterIssuers (or pass --skip-issuers)"
+        )
+
+    issuer_docs = (
+        [_http01_issuer(env, acme_email, settings.ingress_class) for env in ("staging", "prod")]
+        if not skip_issuers and acme_email
+        else []
+    )
+    reusing_cert = not cluster_issuer and host == settings.proxy_host
+    middleware = _build_ipallowlist_middleware(
+        namespace=settings.namespace, source_ranges=source_ranges, xff_depth=xff_depth
+    )
+    ingress = _build_federation_ingress(
+        namespace=settings.namespace,
+        service_name=settings.service_name,
+        port=settings.port,
+        host=host,
+        ingress_class=settings.ingress_class,
+        tls_secret=tls_secret,
+        cluster_issuer=cluster_issuer,
+    )
+
+    click.secho(f"==> Plan (CoreWeave network stack for {settings.cluster}):", fg="blue", bold=True)
+    if not skip_traefik:
+        click.echo(
+            f"  helm upgrade --install {settings.traefik_release} {TRAEFIK_CHART} "
+            f"-n {settings.traefik_namespace} --create-namespace"
+        )
+    if not skip_cert_manager:
+        click.echo(
+            f"  helm upgrade --install {settings.cert_manager_release} {CERT_MANAGER_CHART} "
+            f"-n {settings.cert_manager_namespace} --create-namespace"
+        )
+    if issuer_docs:
+        click.secho("==> ClusterIssuers (HTTP-01 via Traefik):", fg="blue", bold=True)
+        click.echo(yaml.safe_dump_all(issuer_docs, default_flow_style=False, sort_keys=False))
+    click.secho("==> Federation ingress:", fg="blue", bold=True)
+    click.echo(f"  host:         {host}  (CNAME -> the Traefik LoadBalancer FQDN)")
+    click.echo(f"  allowlist:    {', '.join(source_ranges)}")
+    click.echo(f"  backend:      {settings.service_name}:{settings.port} (namespace {settings.namespace})")
+    if reusing_cert:
+        click.echo(f"  tls secret:   {tls_secret}  (reusing the existing /proxy cert on this host)")
+    else:
+        click.echo(f"  tls secret:   {tls_secret}  (cert-manager issues via {cluster_issuer})")
+    if settings.proxy_host == host:
+        click.echo("  note:         /proxy keeps its own world-open router; this allowlists the rest.")
+    click.echo(yaml.safe_dump_all([middleware, ingress], default_flow_style=False, sort_keys=False))
+
+    _warn_if_permissive(settings)
+
+    if not apply:
+        click.secho("\nwarn: dry run — nothing applied. Re-run with --apply to install.", fg="yellow", err=True)
+        return
+
+    click.secho(f"==> Adding/updating helm repo {CW_REPO_NAME} ({CW_REPO_URL})", fg="blue", bold=True)
+    run(["helm", "repo", "add", CW_REPO_NAME, CW_REPO_URL], check=True, stdout=subprocess.DEVNULL)
+    run(["helm", "repo", "update", CW_REPO_NAME], check=True, stdout=subprocess.DEVNULL)
+
+    if not skip_traefik:
+        click.secho(
+            f"==> Installing Traefik ({TRAEFIK_CHART}) in namespace {settings.traefik_namespace}", fg="blue", bold=True
+        )
+        _helm_install(TRAEFIK_CHART, settings.traefik_release, settings.traefik_namespace, traefik_version, hflags)
+
+    if not skip_cert_manager:
+        click.secho(
+            f"==> Installing cert-manager ({CERT_MANAGER_CHART}) in namespace {settings.cert_manager_namespace}",
+            fg="blue",
+            bold=True,
+        )
+        _helm_install(
+            CERT_MANAGER_CHART,
+            settings.cert_manager_release,
+            settings.cert_manager_namespace,
+            cert_manager_version,
+            hflags,
+        )
+
+    if issuer_docs:
+        click.secho("==> Waiting for the cert-manager ClusterIssuer CRD, then applying issuers", fg="blue", bold=True)
+        wait_for_crd(CLUSTERISSUER_CRD, kflags)
+        kubectl_apply_docs(issuer_docs, kflags, "ClusterIssuers")
+
+    click.secho("==> Applying the federation ingress", fg="blue", bold=True)
+    if not resource_present("crd", _MIDDLEWARE_CRD, kflags):
+        raise click.ClickException(
+            f"Traefik Middleware CRD {_MIDDLEWARE_CRD} not found — the ipAllowList cannot be applied.\n"
+            "Install Traefik (drop --skip-traefik) before the federation route."
+        )
+    kubectl_apply_docs([middleware, ingress], kflags, "the federation ingress")
+
+    _print_next_steps(settings, host=host, cluster_issuer=cluster_issuer)
+
+
+def _helm_install(chart: str, release: str, namespace: str, version: str | None, hflags: list[str]) -> None:
+    cmd = ["helm", *hflags, "upgrade", "--install", release, chart, "-n", namespace, "--create-namespace", "--wait"]
+    if version:
+        cmd += ["--version", version]
+    result = run(cmd)
+    if result.returncode != 0:
+        raise click.ClickException(f"helm install of {chart} failed")
+
+
+def _print_next_steps(settings: NetworkSettings, *, host: str, cluster_issuer: str) -> None:
+    """Print exactly what the operator must do next — DNS, controller wiring, verification."""
+    kflags = kubectl_flags(settings.kubeconfig, settings.context)
+    click.secho("==> Done. To finish wiring the CoreWeave network:", fg="green", bold=True)
+
+    # 1. DNS. The host serves both /proxy (if configured) and the federation route.
+    if host == settings.proxy_host:
+        click.secho(f"  1) {host} already resolves (the existing /proxy host) — no new DNS needed.", fg="green")
+    else:
+        click.secho(f"  1) CNAME {host} at the Traefik LoadBalancer FQDN:", fg="green", bold=True)
+        fqdn = read_traefik_fqdn(settings.traefik_release, settings.traefik_namespace, kflags)
+        if fqdn:
+            click.echo(f"        {host}   CNAME   {fqdn}")
+        else:
+            click.secho(
+                "       (Traefik's FQDN isn't allocated yet — read it in a minute, then CNAME to it:)", fg="yellow"
+            )
+            click.echo(
+                f"       kubectl get svc {settings.traefik_release} -n {settings.traefik_namespace} "
+                "-o=jsonpath='{.status.conditions[?(@.type==\"ExternalRecords\")].message}'"
+            )
+            click.echo(f"        {host}   CNAME   <that>{_COREWEAVE_APP}")
+
+    # 2. The controller creates its own /proxy Ingress at start-up; a cluster that
+    # publishes /proxy needs a running controller for that route to exist.
+    if settings.proxy_host:
+        click.secho(
+            f"  2) If {settings.cluster}'s controller has not yet created its /proxy Ingress, restart it so it does:",
+            fg="green",
+        )
+        click.echo(f"       iris --cluster {settings.cluster} cluster restart")
+
+    click.secho("  3) Verify federation reach from the marin controller VM:", fg="green", bold=True)
+    click.echo("       - ListBackends WITH the federation JWT from the allowlisted egress IP  -> succeeds")
+    click.echo("       - the same call WITHOUT the JWT (controller enforcing)                 -> UNAUTHENTICATED")
+    click.echo("       - the same call from a non-allowlisted IP                              -> refused (403)")
+    if cluster_issuer and "staging" in cluster_issuer:
+        click.secho(
+            "  4) Staging issuer in use — once the cert validates, re-run with --skip-traefik "
+            "--skip-cert-manager --skip-issuers --cluster-issuer letsencrypt-http01-prod --apply.",
+            fg="green",
+        )
+    if settings.auth_mode == _NULL_AUTH:
+        click.secho(
+            "  NOTE: this controller is still permissive — add auth.trusted_cidrs and restart so an "
+            "off-cluster request must present a bearer; the IP allowlist alone is not an identity gate.",
+            fg="yellow",
+        )
+    click.echo(f"  (inspect: kubectl get ingress {_INGRESS_NAME} -n {settings.namespace} {' '.join(kflags)} -o wide)")
+
+
+# --------------------------------------------------------------------------
+# Uninstall
+# --------------------------------------------------------------------------
+# kubectl's error for a kind the API server doesn't know (e.g. a CRD already
+# deleted); from k8s.io/cli-runtime's resource builder: "the server doesn't
+# have a resource type %q". Any other nonzero exit (auth, network, …) is a
+# real failure and must not read as "nothing left".
+_UNKNOWN_KIND_MARKER = "doesn't have a resource type"
+
+
+def _resource_names(kind: str, kflags: list[str], selector: str | None = None) -> list[str]:
+    """List ``kind/name`` strings for ``kind``; [] when the kind itself is unknown (CRD gone)."""
+    cmd = ["kubectl", *kflags, "get", kind, "-o", "name"]
+    if selector:
+        cmd += ["-l", selector]
+    result = subprocess.run(cmd, capture_output=True, text=True)
+    if result.returncode == 0:
+        return result.stdout.split()
+    if _UNKNOWN_KIND_MARKER in result.stderr:
+        return []
+    raise click.ClickException(f"`{' '.join(cmd)}` failed: {result.stderr.strip()}")
+
+
+def _short_name(qualified: str) -> str:
+    """``customresourcedefinition.apiextensions.k8s.io/foo.cert-manager.io`` -> ``foo.cert-manager.io``."""
+    return qualified.rsplit("/", 1)[-1]
+
+
+def _group_crds(kflags: list[str]) -> list[str]:
+    """CRDs belonging to the chart-registered API groups (``CRD_GROUPS``)."""
+    suffixes = tuple(f".{group}" for group in CRD_GROUPS)
+    return [crd for crd in _resource_names("crd", kflags) if _short_name(crd).endswith(suffixes)]
+
+
+def _release_scoped(kind: str, releases: tuple[str, ...], kflags: list[str]) -> list[str]:
+    """Cluster-scoped ``kind`` objects belonging to ``releases``.
+
+    Matches by the ``app.kubernetes.io/instance`` label where the charts set it,
+    plus a name-prefix fallback for objects created without instance labels.
+    """
+    found: set[str] = set()
+    all_names = _resource_names(kind, kflags)
+    for release in releases:
+        found.update(_resource_names(kind, kflags, selector=f"{INSTANCE_LABEL}={release}"))
+        found.update(name for name in all_names if _short_name(name).startswith(release))
+    return sorted(found)
+
+
+def _kubectl_delete(names: list[str], kflags: list[str]) -> None:
+    run(["kubectl", *kflags, "delete", *names, "--ignore-not-found"], check=True)
+
+
+def _helm_release_installed(release: str, namespace: str, hflags: list[str]) -> bool:
+    result = subprocess.run(
+        ["helm", *hflags, "status", release, "-n", namespace],
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.DEVNULL,
+    )
+    return result.returncode == 0
+
+
+def _delete_federation_ingress(settings: NetworkSettings, kflags: list[str]) -> None:
+    """Delete the federation Ingress + Middleware in the controller namespace.
+
+    Both live in the controller namespace, not the chart namespaces, so the stack
+    teardown below would not reach the Ingress; the Middleware is a traefik.io CR
+    that the CRD sweep would remove, but deleting it explicitly keeps teardown
+    order-independent.
+    """
+    run(
+        ["kubectl", *kflags, "delete", "ingress", _INGRESS_NAME, "-n", settings.namespace, "--ignore-not-found"],
+        check=True,
+    )
+    if resource_present("crd", _MIDDLEWARE_CRD, kflags):
+        run(
+            [
+                "kubectl",
+                *kflags,
+                "delete",
+                _MIDDLEWARE_CRD,
+                _MIDDLEWARE_NAME,
+                "-n",
+                settings.namespace,
+                "--ignore-not-found",
+            ],
+            check=True,
+        )
+
+
+def uninstall(settings: NetworkSettings, *, apply: bool) -> None:
+    hflags = helm_flags(settings.kubeconfig, settings.context)
+    kflags = kubectl_flags(settings.kubeconfig, settings.context)
+    ingress_class = settings.ingress_class
+    release_pairs = (
+        (settings.traefik_release, settings.traefik_namespace),
+        (settings.cert_manager_release, settings.cert_manager_namespace),
+    )
+    releases = (settings.traefik_release, settings.cert_manager_release)
+    issuers = [ISSUER_NAMES[env] for env in ("staging", "prod")]
+    namespaces = list(dict.fromkeys([settings.traefik_namespace, settings.cert_manager_namespace]))
+
+    click.secho("==> Plan (teardown of the CoreWeave network stack):", fg="blue", bold=True)
+    click.echo(f"  kubectl delete ingress {_INGRESS_NAME} + middleware {_MIDDLEWARE_NAME} -n {settings.namespace}")
+    click.echo(f"  kubectl delete clusterissuer {' '.join(issuers)}   # first, while the CRD still exists")
+    for release, namespace in release_pairs:
+        click.echo(f"  helm uninstall {release} -n {namespace}")
+    click.echo(f"  kubectl delete namespace {' '.join(namespaces)}")
+    groups = " / ".join(f"*.{group}" for group in CRD_GROUPS)
+    click.echo(f"  kubectl delete crd <every {groups} CRD>")
+    click.echo(
+        f"  kubectl delete <{', '.join(SWEEP_KINDS)} with "
+        f"{INSTANCE_LABEL} in ({', '.join(releases)}) or name-prefixed by a release>"
+    )
+    click.echo(f"  kubectl delete ingressclass {ingress_class}")
+    click.echo("  then verify nothing is left")
+
+    if not apply:
+        click.secho("\nwarn: dry run — nothing deleted. Re-run with --apply to uninstall.", fg="yellow", err=True)
+        return
+
+    # 0. The federation route (in the controller namespace, outside the chart teardown).
+    click.secho("==> Deleting the federation ingress", fg="blue", bold=True)
+    _delete_federation_ingress(settings, kflags)
+
+    # 1. ClusterIssuers next: once the CRD is deleted they become unaddressable.
+    click.secho("==> Deleting ClusterIssuers (before the CRD goes)", fg="blue", bold=True)
+    if CLUSTERISSUER_CRD in {_short_name(crd) for crd in _resource_names("crd", kflags)}:
+        present = {_short_name(name) for name in _resource_names("clusterissuer", kflags)}
+        for name in (issuer for issuer in issuers if issuer not in present):
+            click.echo(f"  clusterissuer/{name}: already absent")
+        found = [f"clusterissuer/{issuer}" for issuer in issuers if issuer in present]
+        if found:
+            _kubectl_delete(found, kflags)
+    else:
+        click.echo(f"  CRD {CLUSTERISSUER_CRD} already absent — no ClusterIssuers to delete")
+
+    # 2. helm uninstall both releases (helm removes its own cluster-scoped objects).
+    for release, namespace in release_pairs:
+        click.secho(f"==> Uninstalling helm release {release} (namespace {namespace})", fg="blue", bold=True)
+        if not _helm_release_installed(release, namespace, hflags):
+            click.echo(f"  release {release} not installed in {namespace} — already absent")
+            continue
+        result = run(["helm", *hflags, "uninstall", release, "-n", namespace, "--wait"])
+        if result.returncode != 0:
+            raise click.ClickException(f"helm uninstall of {release} failed")
+
+    # 3. The namespaces the installs created.
+    click.secho("==> Deleting namespaces", fg="blue", bold=True)
+    existing = {_short_name(name) for name in _resource_names("namespace", kflags)}
+    for namespace in (ns for ns in namespaces if ns not in existing):
+        click.echo(f"  namespace/{namespace}: already absent")
+    found = [f"namespace/{ns}" for ns in namespaces if ns in existing]
+    if found:
+        _kubectl_delete(found, kflags)
+
+    # 4. Sweep the cluster-scoped leftovers a namespace delete orphans.
+    click.secho("==> Sweeping cluster-scoped leftovers", fg="blue", bold=True)
+    crds = _group_crds(kflags)
+    if crds:
+        _kubectl_delete(crds, kflags)
+    else:
+        click.echo(f"  CRDs ({groups}): already absent")
+    for kind in SWEEP_KINDS:
+        names = _release_scoped(kind, releases, kflags)
+        if names:
+            _kubectl_delete(names, kflags)
+        else:
+            click.echo(f"  {kind}: already absent")
+    if any(_short_name(name) == ingress_class for name in _resource_names("ingressclass", kflags)):
+        _kubectl_delete([f"ingressclass/{ingress_class}"], kflags)
+    else:
+        click.echo(f"  ingressclass/{ingress_class}: already absent")
+
+    # 5. Verify: list anything still present rather than assuming the sweeps worked.
+    click.secho("==> Verifying teardown", fg="blue", bold=True)
+    leftovers: list[str] = []
+    fed_ingress = subprocess.run(
+        ["kubectl", *kflags, "get", "ingress", _INGRESS_NAME, "-n", settings.namespace],
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.DEVNULL,
+    )
+    if fed_ingress.returncode == 0:
+        leftovers.append(f"ingress/{_INGRESS_NAME} (namespace {settings.namespace})")
+    for release, namespace in release_pairs:
+        if _helm_release_installed(release, namespace, hflags):
+            leftovers.append(f"helm release {release} (namespace {namespace})")
+    existing = {_short_name(name) for name in _resource_names("namespace", kflags)}
+    leftovers += [f"namespace/{ns}" for ns in namespaces if ns in existing]
+    leftovers += _group_crds(kflags)
+    for kind in SWEEP_KINDS:
+        leftovers += _release_scoped(kind, releases, kflags)
+    leftovers += [
+        f"ingressclass/{ingress_class}"
+        for name in _resource_names("ingressclass", kflags)
+        if _short_name(name) == ingress_class
+    ]
+    if leftovers:
+        click.secho("warn: still present after teardown:", fg="yellow", err=True)
+        for item in leftovers:
+            click.echo(f"  {item}", err=True)
+        raise click.ClickException("teardown incomplete — delete the resources above by hand")
+    click.secho(
+        "==> Teardown complete — no releases, namespaces, or cluster-scoped leftovers remain.", fg="green", bold=True
+    )
+
+
+# --------------------------------------------------------------------------
+# CLI
+# --------------------------------------------------------------------------
+@click.group()
+@click.option(
+    "--cluster",
+    required=True,
+    help="Iris cluster name; supplies host, ingress class, namespace, controller Service, and kubeconfig.",
+)
+@click.option("--kubeconfig", default="", help="kubeconfig to use [default: the cluster's].")
+@click.option("--context", default="", help="kube context to target [default: the cluster's kube_context].")
+@click.option("--traefik-namespace", default="traefik", show_default=True)
+@click.option("--traefik-release", default="traefik", show_default=True, help="helm release name for Traefik.")
+@click.option("--cert-manager-namespace", default="cert-manager", show_default=True)
+@click.option(
+    "--cert-manager-release", default="cert-manager", show_default=True, help="helm release name for cert-manager."
+)
+@click.pass_context
+def main(
+    ctx: click.Context,
+    cluster: str,
+    kubeconfig: str,
+    context: str,
+    traefik_namespace: str,
+    traefik_release: str,
+    cert_manager_namespace: str,
+    cert_manager_release: str,
+) -> None:
+    """Traefik + cert-manager + HTTP-01 issuers + the IP-locked federation ingress for a CoreWeave cluster."""
+    derived = _derive_from_cluster(cluster)
+    ctx.obj = NetworkSettings(
+        cluster=cluster,
+        namespace=derived.namespace,
+        service_name=derived.service_name,
+        port=derived.port,
+        ingress_class=derived.ingress_class,
+        proxy_host=derived.proxy_host,
+        proxy_tls_secret=derived.proxy_tls_secret,
+        auth_mode=derived.auth_mode,
+        kubeconfig=kubeconfig or derived.kubeconfig or None,
+        context=context or derived.context or None,
+        traefik_namespace=traefik_namespace,
+        traefik_release=traefik_release,
+        cert_manager_namespace=cert_manager_namespace,
+        cert_manager_release=cert_manager_release,
+    )
+
+
+@main.command("install")
+@click.option(
+    "--acme-email",
+    default=None,
+    help="Email for the Let's Encrypt HTTP-01 ClusterIssuers (required unless --skip-issuers).",
+)
+@click.option(
+    "--allow-source",
+    "allow_sources",
+    multiple=True,
+    required=True,
+    help="IP or CIDR permitted to reach the federation route (repeatable). The marin egress IP.",
+)
+@click.option(
+    "--host",
+    default="",
+    help="Federation ingress host [default: the cluster's /proxy host, else iris-cw-<cluster>.oa.dev].",
+)
+@click.option(
+    "--tls-secret", default="", help="TLS secret [default: reuse the cluster's /proxy secret, else a new one]."
+)
+@click.option(
+    "--cluster-issuer",
+    default="",
+    help="cert-manager ClusterIssuer for the federation cert [default: none when reusing an existing cert, "
+    "else letsencrypt-http01-staging].",
+)
+@click.option(
+    "--xff-depth",
+    default=0,
+    show_default=True,
+    help="X-Forwarded-For depth for the ipAllowList (0 = match the direct peer; >0 if the LB SNATs).",
+)
+@click.option("--traefik-version", default=None, help="Pin the Traefik chart version (default: latest).")
+@click.option("--cert-manager-version", default=None, help="Pin the cert-manager chart version (default: latest).")
+@click.option("--skip-traefik", is_flag=True, help="Do not install Traefik (already present).")
+@click.option("--skip-cert-manager", is_flag=True, help="Do not install cert-manager (already present).")
+@click.option("--skip-issuers", is_flag=True, help="Do not create the HTTP-01 ClusterIssuers (already present).")
+@click.option("--apply/--no-apply", default=False, help="Actually mutate the cluster (default: dry-run only).")
+@click.pass_obj
+def install_cmd(
+    settings: NetworkSettings,
+    acme_email: str | None,
+    allow_sources: tuple[str, ...],
+    host: str,
+    tls_secret: str,
+    cluster_issuer: str,
+    xff_depth: int,
+    traefik_version: str | None,
+    cert_manager_version: str | None,
+    skip_traefik: bool,
+    skip_cert_manager: bool,
+    skip_issuers: bool,
+    apply: bool,
+) -> None:
+    """Install the whole CoreWeave network stack (dry-run without --apply)."""
+    source_ranges = [_normalize_source(value) for value in allow_sources]
+    resolved_host = host or _default_host(settings)
+    # Reuse the controller's existing /proxy cert when publishing on that same
+    # host; otherwise cert-manager issues a fresh cert for the federation route.
+    reuse_existing_cert = resolved_host == settings.proxy_host and bool(settings.proxy_tls_secret)
+    resolved_tls_secret = tls_secret or (settings.proxy_tls_secret if reuse_existing_cert else DEFAULT_TLS_SECRET)
+    resolved_issuer = cluster_issuer or ("" if reuse_existing_cert else DEFAULT_CLUSTER_ISSUER)
+    install(
+        settings,
+        host=resolved_host,
+        source_ranges=source_ranges,
+        tls_secret=resolved_tls_secret,
+        cluster_issuer=resolved_issuer,
+        xff_depth=xff_depth,
+        acme_email=acme_email,
+        traefik_version=traefik_version,
+        cert_manager_version=cert_manager_version,
+        skip_traefik=skip_traefik,
+        skip_cert_manager=skip_cert_manager,
+        skip_issuers=skip_issuers,
+        apply=apply,
+    )
+
+
+@main.command("uninstall")
+@click.option("--apply/--no-apply", default=False, help="Actually mutate the cluster (default: dry-run only).")
+@click.pass_obj
+def uninstall_cmd(settings: NetworkSettings, apply: bool) -> None:
+    """Tear the whole network stack back down and verify nothing remains."""
+    uninstall(settings, apply=apply)
+
+
+if __name__ == "__main__":
+    main()
